Fixed issue when trying to execute module via `python -m flux_led`
Before:
No module named flux_led.__main__; 'flux_led' is a package and cannot be directly executed

After:
$ python3 -m flux_led
Usage: usage: __main__.py [-sS10cwpCiltThe] [addr1 [addr2 [addr3] ...].

__main__.py: error: An operation must be specified

$ python3 -m flux_led -s
1 bulbs found
  ACCF23920AAA 192.168.1.1
